backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart
httpx
Pillow